backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
web3==6.0.0
python-dotenv==1.0.0
twilio==8.10.0
setuptools==69.0.0
